backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
PyPDF2
python-multipart
faiss-cpu
numpy
